requirements.txt
appdirs==1.4.3
asgiref==3.2.5
attrs==19.3.0
distlib==0.3.0
Django==3.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tox==3.14.5
uWSGI==2.0.18
virtualenv==20.0.12
wcwidth==0.1.8
zipp==3.1.0
